+++ src/app/core/ase-footer/ase-footer.component.ts
@@ -1,0 +1,12 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'ase-footer',
+  standalone: true,
+  imports: [],
+  templateUrl: './ase-footer.component.html',
+  styleUrl: './ase-footer.component.scss'
+})
+export class AseFooterComponent {
+
+}